# swagger.yaml
openapi: 3.0.0

info:
  title: Task API Auth
  version: 1.0.0
  description: Authentication API using Node.js, Express and Supabase

servers:
  - url: http://localhost:3000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:

  /auth/signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        "201":
          description: User created
        "400":
          description: Invalid input

  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        "200":
          description: Login successful
        "400":
          description: Missing email or password
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logout successful
        "401":
          description: Unauthorized

  /public/info:
    get:
      summary: Public information
      responses:
        "200":
          description: Public information

  /protected/profile:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
        "401":
          description: Invalid or expired token

  /protected/dashboard:
    get:
      summary: Protected dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard information
        "401":
          description: Invalid or expired token